+++ src/pages/placeDetails/PlaceDetails.tsx
@@ -14,21 +14,13 @@ import { useParams, useNavigate } from "react-router-dom"
 import {getPlaceDetailsInfo} from '../../api'
 import {PlaceBoardData} from '../../data/placeSearch'
 import {faList} from '@fortawesome/free-solid-svg-icons'
-import { useSelector } from 'react-redux';
-import { RootState } from '../../store/rootReducer';
-
 
 // 장소 상세 페이지
-
-
 export const PlaceDetails = () => {
 
     const [boardData, setBoardData] = useState<PlaceBoardData[] | null>(null);
     const { pno } = useParams();
-    const navigate = useNavigate();
-
-    const user = useSelector((state: RootState) => state.login.mno)!
-
+    
     async function fetchData(pnoParam: string | undefined) {
         const pnoNumber = pnoParam ? Number(pno) : undefined; // pno를 숫자로 변환
         try {
@@ -40,31 +32,20 @@ export const PlaceDetails = () => {
         } catch (err) {
             console.log(err);
             alert('게시글이 없습니다!');
-            navigate(-1)
         }
     }
-
 
     useEffect(() => {
         fetchData(pno); // fetchData 함수를 실행하여 초기 데이터를 가져옴
     }, [pno]); // 빈 배열을 전달하여 컴포넌트가 처음 마운트될 때만 실행
 
-    const handleRegisterClick = () => {
-            navigate(`/board/place/posting/register`);
-        }
-
-
     return (
         <Box>
-            <div className='flex justify-end w-4/6 '>
-               {user && <Button onClick={handleRegisterClick} className="h-16 text-xl text-white w-36 bg-darkGreen" value={'게시글 작성'}/> } 
-            </div>
             <div className="flex justify-center w-full">
-                <div className="w-2/3">
-                    
-                        {boardData && (
-                            <SearchMap places={boardData} className="w-full h-full" innerRef={null} />
-                        )}
+                <div className="w-2/3 m-10 rounded-sm">     
+                    {boardData && (
+                        <SearchMap places={boardData} className="w-full h-full" innerRef={null} />
+                    )}
                 </div>
             </div>
             <BoardToggle>
